requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
gigachat
requests
fastapi
pydantic